Restore corrupted pi-coding-agent manifest; recover misplaced functions into messages.rs

+++ crates/pi-tui/src/components/cancellable_loader.rs
@@ -58,7 +58,7 @@ impl CancellableLoader {
 }
 
 impl Component for CancellableLoader {
-    fn render(&mut self, width: usize) -> Vec<String> {
+    fn render(&mut self, width: f64) -> Vec<String> {
         self.loader.render(width)
     }
